Lot of changes:
Bytes.ParseSequence renamed to Bytes.ParseSubBytes
It's more reflecting the reality: we are parsing a sub-part of the byte slice, but not always a "sequence"

New type of functions: "ReadTagged..."
It is similar to the "Read..." functions but it gets two additional parameters "class" and "tag".
The fact is that we often parse some types (for example OCTETSTRING) which universal tag if hidden by a context specific tag.
The "Read" functions still exist but will call the "ReadTagged" function with a "universal" class and default type tag value.

New error type: LdapError

Error handling: concatenate the name of the current function when forwarding an error, thus building a stack trace.
This will have to improved. Lot of places still have the default behavior "if err != nil { return err }"

New type of method: "Read...Components"
In the ASN1 vocabulary, "COMPONENTS OF" means the fields of a structure.
When calling "Bytes.ParseSubBytes" we call a "Read...Components" method which will fill the object attributes.

New request parsing: the SearchRequest !
That was hard, it's a very complex type. All the previous listed changes are a consequence of this implementation.
Currently only one unit test is present, I'll add some more when testing.